Reorganize structure of filter modules
With all of the modules in the same file it masks how many modules there
are and forces explicit ordering within the file. This extracts each of
the modules into its own file and uses namespacing to indicate the
dependency structure.
