+++ mix.exs
@@ -27,8 +27,8 @@ defmodule Honeybadger.Mixfile do
            proxy: nil,
            proxy_auth: {nil, nil},
            use_logger: true,
-           notice_filter: Honeybadger.DefaultNoticeFilter,
-           filter: Honeybadger.DefaultFilter,
+           notice_filter: Honeybadger.NoticeFilter.Default,
+           filter: Honeybadger.Filter.Default,
            filter_keys: [:password, :credit_card],
            filter_disable_url: false,
            filter_disable_params: false,
